Further hacking on Tree functionality.
We now use pathlib.Path objects underneath the hood for Trees and
Leaves. We don't use them directly instead of Trees and Leaves or as
base classes because we honestly don't need all Path's functionality at
that level. Well, that may change, but for now this works well. Still
experimenting.